# src/openapi.yaml
# yaml-language-server: $schema=https://raw.githubusercontent.com/OAI/OpenAPI-Specification/main/schemas/v3.0/schema.json
openapi: 3.0.3
info:
    description: |-
        This is API Documentation for Patungin.id
    version: '0.1'
    title: Patunginid - 0.1

paths:
    /checkhealth:
        get:
            description: Check health api connection
            responses:
                200:
                    description: Successfully returns a message
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    status:
                                        type: number
                                        example: 200
                                    data:
                                        type: object
                                        properties:
                                            message:
                                                type: string
                                                example: 'Negative'

    /checkhealth-db:
        get:
            description: Check health api db connection
            responses:
                200:
                    description: Successfully returns a message
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    status:
                                        type: number
                                        example: 200
                                    data:
                                        type: object
                                        properties:
                                            message:
                                                type: number
                                                example: 4

    /api/v1/receipt/calculate:
        post:
            description: calculate prices and distribute it accordingly
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                title:
                                    description: the transaction title
                                    type: string
                                users:
                                    type: array
                                    items:
                                        type: string
                                items:
                                    type: array
                                    items:
                                        type: object
                                        properties:
                                            name:
                                                type: string
                                            total_price:
                                                type: number
                                            payer:
                                                description: an array of index of users that pays
                                                type: array
                                                items:
                                                    type: integer
                                subtotal:
                                    description: subtotal (sum of items price)
                                    type: number
                                tax_price:
                                    type: number
                                service_charge:
                                    type: number
                                other_price:
                                    type: number
                                discount:
                                    type: number
                                total_price:
                                    type: number
            responses:
                200:
                    description: Successfully distributed the prices
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    status:
                                        type: number
                                        example: 200
                                    data:
                                        type: object
                                        properties:
                                            id:
                                                type: string
                                            title:
                                                type: string
                                            users:
                                                type: array
                                                items:
                                                    type: object
                                                    properties:
                                                        id:
                                                            type: string
                                                        name:
                                                            type: string
                                                        items:
                                                            type: array
                                                            items:
                                                                type: array
                                                                items:
                                                                    type: object
                                                                    properties:
                                                                        name:
                                                                            type: string
                                                                        price:
                                                                            type: number
                                                        tax_price:
                                                            type: number
                                                        service_charge:
                                                            type: number
                                                        other_price:
                                                            type: number
                                                        discount:
                                                            type: number
                                                        total_price:
                                                            type: number

    /api/v1/receipt/{uuid}:
        get:
            description: get a receipt detail that has been splited
            parameters:
                - name: uuid
                  in: path
                  description: uuid of the receipt
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: Successfully return the receipt and details
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    status:
                                        type: number
                                        example: 200
                                    data:
                                        type: object
                                        properties:
                                            receipt:
                                                type: object
                                                properties:
                                                    id:
                                                        type: string
                                                    title:
                                                        type: string
                                                    tax_price:
                                                        type: number
                                                    service_charge:
                                                        type: number
                                                    other_price:
                                                        type: number
                                                    discount:
                                                        type: number
                                                    total_price:
                                                        type: number
                                                    payers:
                                                        type: array
                                                        items:
                                                            type: object
                                                            properties:
                                                                id:
                                                                    type: string
                                                                name:
                                                                    type: string
                                                                items:
                                                                    type: array
                                                                    items:
                                                                        type: array
                                                                        items:
                                                                            type: object
                                                                            properties:
                                                                                id:
                                                                                    type: string
                                                                                name:
                                                                                    type: string
                                                                                price:
                                                                                    type: number
                                                                tax_price:
                                                                    type: number
                                                                service_charge:
                                                                    type: number
                                                                other_price:
                                                                    type: number
                                                                discount:
                                                                    type: number
                                                                total_price:
                                                                    type: number

    /api/v1/receipt/{uuid}/pay/{user_id}/cva:
        post:
            description: user id payment with cva
            parameters:
                - name: uuid
                  in: path
                  description: uuid of the receipt
                  required: true
                  schema:
                      type: string
                - name: user_id
                  in: path
                  description: uuid of the user in the transaction
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                bankShortCode:
                                    type: string
                                destination_account_name:
                                    type: string
                                destination_account_number:
                                    type: string
                                destination_bank_code:
                                    type: string
            responses:
                200:
                    description: Successfully created the cva
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    success:
                                        type: boolean
                                        example: true
                                    status:
                                        type: number
                                        example: 200
                                    data:
                                        type: object
                                        properties:
                                            va_account:
                                                type: string
                                            display_name:
                                                type: string
                                            description:
                                                type: string
                                            destination_account_name:
                                                type: string
                                            destination_account_number:
                                                type: string
                                            destination_bank_code:
                                                type: string
                                            status:
                                                type: string
                                            total_price:
                                                type: string

components:
    schemas:
        ApiResponse:
            type: object
            properties:
                success:
                    type: boolean
                status:
                    type: number

    securitySchemes:
        api_key:
            type: apiKey
            name: api_key
            in: header